Done with Navbar and Footer Ticket
Footer should always be at bottom, NavBar looks close to Figma, and we put in some things to help with different screen sizes.

Made some changes to page.tsx so that we can show the navbar and footer and get rid of placeholder content.

Added just some pictures from the figma for the profile placeholder pictures