Done with Navbar and Footer Ticket (#9)
* Done with Navbar and Footer Ticket

Footer should always be at bottom, NavBar looks close to Figma, and we put in some things to help with different screen sizes.

Made some changes to page.tsx so that we can show the navbar and footer and get rid of placeholder content.

Added just some pictures from the figma for the profile placeholder pictures

* Changes Finalized for Navbar and Footer

Made changes for consistency in code as well as adding navbar and footer to layout.tsx

Co-Authored-By: priyankaonta <68300973+priyankaonta@users.noreply.github.com>

---------

Co-authored-by: priyankaonta <68300973+priyankaonta@users.noreply.github.com>